chore(frontend): move route reuse strategy to routing folder
